# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Web Project API",
        "description": "API Documentation for Web Project — RMUTR",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/webproject/",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "เข้าสู่ระบบด้วย Email และ Password",
                "description": "1. API เข้าสู่ระบบด้วย Email + Password\n   - ส่ง profile_image จาก DB กลับไปด้วย (null ถ้าไม่มี)",
                "operationId": "222b2625838e3c57e017987f73598fda",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "user@rmutr.ac.th"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "เข้าสู่ระบบสำเร็จ"
                    },
                    "401": {
                        "description": "อีเมลหรือรหัสผ่านไม่ถูกต้อง"
                    },
                    "403": {
                        "description": "กรุณาใช้อีเมลของมหาวิทยาลัย (@rmutr.ac.th) เท่านั้น"
                    }
                }
            }
        },
        "/api/login/google": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "เข้าสู่ระบบด้วย Google Access Token",
                "description": "2. API เข้าสู่ระบบด้วย Google Token\n   - บันทึกรูป Google ขนาด 400px\n   - อัปเดตรูปทุกครั้งที่ Login ด้วย Google",
                "operationId": "882bab19633a0b0ac04983ca02db4e49",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "access_token"
                                ],
                                "properties": {
                                    "access_token": {
                                        "type": "string",
                                        "example": "ya29.a0..."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "เข้าสู่ระบบด้วย Google สำเร็จ"
                    },
                    "401": {
                        "description": "Google Token ไม่ถูกต้อง"
                    },
                    "403": {
                        "description": "กรุณาใช้อีเมลของมหาวิทยาลัย (@rmutr.ac.th) เท่านั้น"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "ออกจากระบบ",
                "description": "3. API ออกจากระบบ",
                "operationId": "ad65cbbd4e9f201619eae184a5961a98",
                "responses": {
                    "200": {
                        "description": "ออกจากระบบสำเร็จ"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth"
        }
    ]
}